OMS rootless: Introducing Signature check in Masquerade (1/3)
This commit introduces a signing Signature check against all AUTHORIZED_CALLERS.
It can be bypassed automatically if the ROM is a userdebug or eng build, or if
the user has explicitly toggled "Developer options" secure setting.

The Signature String was produced with the following call:
log(pi.signatures[0].toCharsString());

and it was deliberately added into Java and not a Resources string so it will
be impossible to overlay.

The build type check was made configurable for OEMs using a prop value:
ro.masquerade.buildtype.check

the ro.* namespace was chosen for better user security (it shouldn't be writable)

It borrows some of the signature check code from the DashClock app:
https://github.com/romannurik/dashclock/commit/a9bf6636f764b150d489d0817ee7f7796af1a155

Change-Id: I6e9bb97de02a9cc89e57fd4099230cd86e820aa9